Expand a time_period vector to the full range of possible times

Derive from a vector of observation time_periods, a complete ordered sequence of periods in a regular time series, where the length of the periods is specified, as a number of days, weeks, years etc. E.g. this can convert a random set of times to a ordered complete list of 1 week intervals (or 2 month intervals) spanning the same range as the dates. This has some interesting problems regarding where to put breaks within a month or week. Often this is either based on a specific date (e.g. yearly periods starting at 2020-01-01) or a day of week (e.g. 2 weekly periods staring on a sunday) or maybe relative to the input time series (weekly ending on the last date of the data). There is also a problem when we consider data that may have incomplete starting and end periods, which may not be comparable to other periods, and we may need to exclude these from the result.

Usage

# S3 method for class 'time_period'
date_seq(x, period = attributes(x)$unit, complete = FALSE, ...)

Arguments

x

a time period vector

period

the gap between observations as a number of days or as a natural language definition of the period such as "1 week", '2 weeks', '1 month', etc. If not given this will be derived from the dates.

complete

truncate incomplete start and end periods

...

ignored

Value

a vector of time_periods for regular periods between the minimum and maximum of dates, with the boundaries defined by the anchor.

Examples

tmp = as.time_period(c(0,10,100), 7, "2020-01-01")
date_seq(tmp, "7 days")
#> time unit: week, origin: 2020-01-01 (a Wednesday)
#> 0 10 20 30 40 50 60 70 80 90 100
date_seq(tmp, "1 day")
#> time unit: week, origin: 2020-01-01 (a Wednesday)
#> 0.14 0.29 0.43 0.57 0.71 0.86 1 1.14 1.29 1.43 1.57 1.71 1.86 2 2.14 2.29 2.43 2.57 2.71 2.86 3 3.14 3.29 3.43 3.57 3.71 3.86 4 4.14 4.29 4.43 4.57 4.71 4.86 5 5.14 5.29 5.43 5.57 5.71 5.86 6 6.14 6.29 6.43 6.57 6.71 6.86 7 7.14 7.29 7.43 7.57 7.71 7.86 8 8.14 8.29 8.43 8.57 8.71 8.86 9 9.14 9.29 9.43 9.57 9.71 9.86 10 10.14 10.29 10.43 10.57 10.71 10.86 11 11.14 11.29 11.43 11.57 11.71 11.86 12 12.14 12.29 12.43 12.57 12.71 12.86 13 13.14 13.29 13.43 13.57 13.71 13.86 14 14.14 14.29 14.43 14.57 14.71 14.86 15 15.14 15.29 15.43 15.57 15.71 15.86 16 16.14 16.29 16.43 16.57 16.71 16.86 17 17.14 17.29 17.43 17.57 17.71 17.86 18 18.14 18.29 18.43 18.57 18.71 18.86 19 19.14 19.29 19.43 19.57 19.71 19.86 20 20.14 20.29 20.43 20.57 20.71 20.86 21 21.14 21.29 21.43 21.57 21.71 21.86 22 22.14 22.29 22.43 22.57 22.71 22.86 23 23.14 23.29 23.43 23.57 23.71 23.86 24 24.14 24.29 24.43 24.57 24.71 24.86 25 25.14 25.29 25.43 25.57 25.71 25.86 26 26.14 26.29 26.43 26.57 26.71 26.86 27 27.14 27.29 27.43 27.57 27.71 27.86 28 28.14 28.29 28.43 28.57 28.71 28.86 29 29.14 29.29 29.43 29.57 29.71 29.86 30 30.14 30.29 30.43 30.57 30.71 30.86 31 31.14 31.29 31.43 31.57 31.71 31.86 32 32.14 32.29 32.43 32.57 32.71 32.86 33 33.14 33.29 33.43 33.57 33.71 33.86 34 34.14 34.29 34.43 34.57 34.71 34.86 35 35.14 35.29 35.43 35.57 35.71 35.86 36 36.14 36.29 36.43 36.57 36.71 36.86 37 37.14 37.29 37.43 37.57 37.71 37.86 38 38.14 38.29 38.43 38.57 38.71 38.86 39 39.14 39.29 39.43 39.57 39.71 39.86 40 40.14 40.29 40.43 40.57 40.71 40.86 41 41.14 41.29 41.43 41.57 41.71 41.86 42 42.14 42.29 42.43 42.57 42.71 42.86 43 43.14 43.29 43.43 43.57 43.71 43.86 44 44.14 44.29 44.43 44.57 44.71 44.86 45 45.14 45.29 45.43 45.57 45.71 45.86 46 46.14 46.29 46.43 46.57 46.71 46.86 47 47.14 47.29 47.43 47.57 47.71 47.86 48 48.14 48.29 48.43 48.57 48.71 48.86 49 49.14 49.29 49.43 49.57 49.71 49.86 50 50.14 50.29 50.43 50.57 50.71 50.86 51 51.14 51.29 51.43 51.57 51.71 51.86 52 52.14 52.29 52.43 52.57 52.71 52.86 53 53.14 53.29 53.43 53.57 53.71 53.86 54 54.14 54.29 54.43 54.57 54.71 54.86 55 55.14 55.29 55.43 55.57 55.71 55.86 56 56.14 56.29 56.43 56.57 56.71 56.86 57 57.14 57.29 57.43 57.57 57.71 57.86 58 58.14 58.29 58.43 58.57 58.71 58.86 59 59.14 59.29 59.43 59.57 59.71 59.86 60 60.14 60.29 60.43 60.57 60.71 60.86 61 61.14 61.29 61.43 61.57 61.71 61.86 62 62.14 62.29 62.43 62.57 62.71 62.86 63 63.14 63.29 63.43 63.57 63.71 63.86 64 64.14 64.29 64.43 64.57 64.71 64.86 65 65.14 65.29 65.43 65.57 65.71 65.86 66 66.14 66.29 66.43 66.57 66.71 66.86 67 67.14 67.29 67.43 67.57 67.71 67.86 68 68.14 68.29 68.43 68.57 68.71 68.86 69 69.14 69.29 69.43 69.57 69.71 69.86 70 70.14 70.29 70.43 70.57 70.71 70.86 71 71.14 71.29 71.43 71.57 71.71 71.86 72 72.14 72.29 72.43 72.57 72.71 72.86 73 73.14 73.29 73.43 73.57 73.71 73.86 74 74.14 74.29 74.43 74.57 74.71 74.86 75 75.14 75.29 75.43 75.57 75.71 75.86 76 76.14 76.29 76.43 76.57 76.71 76.86 77 77.14 77.29 77.43 77.57 77.71 77.86 78 78.14 78.29 78.43 78.57 78.71 78.86 79 79.14 79.29 79.43 79.57 79.71 79.86 80 80.14 80.29 80.43 80.57 80.71 80.86 81 81.14 81.29 81.43 81.57 81.71 81.86 82 82.14 82.29 82.43 82.57 82.71 82.86 83 83.14 83.29 83.43 83.57 83.71 83.86 84 84.14 84.29 84.43 84.57 84.71 84.86 85 85.14 85.29 85.43 85.57 85.71 85.86 86 86.14 86.29 86.43 86.57 86.71 86.86 87 87.14 87.29 87.43 87.57 87.71 87.86 88 88.14 88.29 88.43 88.57 88.71 88.86 89 89.14 89.29 89.43 89.57 89.71 89.86 90 90.14 90.29 90.43 90.57 90.71 90.86 91 91.14 91.29 91.43 91.57 91.71 91.86 92 92.14 92.29 92.43 92.57 92.71 92.86 93 93.14 93.29 93.43 93.57 93.71 93.86 94 94.14 94.29 94.43 94.57 94.71 94.86 95 95.14 95.29 95.43 95.57 95.71 95.86 96 96.14 96.29 96.43 96.57 96.71 96.86 97 97.14 97.29 97.43 97.57 97.71 97.86 98 98.14 98.29 98.43 98.57 98.71 98.86 99 99.14 99.29 99.43 99.57 99.71 99.86