Summarise a line list with daily case ascertainment variability

This function introduces additional noise into a case series by ensuing a random proportion of the cases are detected each day. The randomness is defined by a logit-normal distribution parametrised by baseline probability p.asc and day to day variability in terms of a dispersion parameter kappa.asc.

Usage

.summarise_with_ascertainment(df, p.asc = 1, kappa.asc = 1)

Arguments

df

the line list

p.asc

the background probability of a case being detected

kappa.asc

a dispersion parameter controlling the day to day variability of ascertainment between 0 (no dispersion) and Inf (maximum)

Value

a count data frame.

Examples

if (interactive()) {

tmp = .test_branching_process(
  changes = tibble::tibble(t = c(0,20,40,60,80,110), R_t = c(1.8,1.5,0.9,1.5,0.8,1.2)),
  kappa = 2,
  max_time = 120,
  p.asc= 0.8,
  kappa.asc = 1.01,
  seed = 100,
  summarise = FALSE
)

tmp2 = dplyr::bind_rows(lapply(c(0,0.5,1,2), function(d) {
  tmp %>% .summarise_with_ascertainment(0.7,d) %>% dplyr::mutate(kappa = d)
}))

ggplot2::ggplot(
  tmp2,
  ggplot2::aes(x=time,y=count, colour=as.factor(kappa))
)+
ggplot2::geom_line()+
ggplot2::geom_point()

}