Calculate a normalised incidence rate per capita

This assumes positive disease counts are stratified by a population grouping, e.g. geography or age, and we have estimates of the size of that population during that time period. Normalising by population size allows us to compare groups.

Usage

normalise_incidence(
  modelled = i_incidence_model,
  pop = i_population_data,
  ...,
  population_unit = 1e+05,
  normalise_time = FALSE
)

Arguments

modelled

Model output from processing the raw dataframe with something like poission_locfit_model - a dataframe with columns:

time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`

incidence.fit (double) - an estimate of the incidence rate on a log scale

incidence.se.fit (positive_double) - the standard error of the incidence rate estimate on a log scale

incidence.0.025 (positive_double) - lower confidence limit of the incidence rate (true scale)

incidence.0.5 (positive_double) - median estimate of the incidence rate (true scale)

incidence.0.975 (positive_double) - upper confidence limit of the incidence rate (true scale)

Any grouping allowed.

pop

The population data must be grouped in the same way as modelled. - a dataframe with columns:

population (positive_integer) - Size of population

Any grouping allowed.

...

not used

population_unit

what population unit do you want the incidence in e.g. per 100K

normalise_time

The default behaviour for incidence is to keep it in the same time units as the input data. If this parameter is set to TRUE the incidence rates are calculated per year. If given as a lubridate period string e.g. "1 day" then the incidence is calculated over that time period.

Value

a dataframe with incidence rates per unit capita. A dataframe containing the following columns:

time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a time_period

incidence.per_capita.fit (double) - an estimate of the incidence per capita rate on a log scale

incidence.per_capita.se.fit (positive_double) - the standard error of the incidence per capita rate estimate on a log scale

incidence.per_capita.0.025 (positive_double) - lower confidence limit of the incidence per capita rate (true scale)

incidence.per_capita.0.5 (positive_double) - median estimate of the incidence per capita rate (true scale)

incidence.per_capita.0.975 (positive_double) - upper confidence limit of the incidence per capita rate (true scale)

population_unit (double) - The population unit on which the per capita incidence rate is calculated

time_unit (lubridate::as.period) - The time period over which the per capita incidence rate is calculated

Any grouping allowed.

Examples

ggoutbreak::england_covid_poisson_age_stratified %>%
  ggoutbreak::normalise_incidence(ggoutbreak::england_demographics) %>%
  dplyr::glimpse()
#> Rows: 26,790
#> Columns: 28
#> Groups: class [19]
#> $ class                       <fct> 00_04, 00_04, 00_04, 00_04, 00_04, 00_04, …
#> $ time                        <time_prd> 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11,…
#> $ incidence.fit               <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ incidence.se.fit            <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ incidence.0.025             <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ incidence.0.05              <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ incidence.0.25              <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ incidence.0.5               <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ incidence.0.75              <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ incidence.0.95              <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ incidence.0.975             <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ growth.fit                  <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ growth.se.fit               <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ growth.0.025                <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ growth.0.05                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ growth.0.25                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ growth.0.5                  <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ growth.0.75                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ growth.0.95                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ growth.0.975                <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ population                  <int> 3077000, 3077000, 3077000, 3077000, 307700…
#> $ incidence.per_capita.0.025  <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ incidence.per_capita.0.5    <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ incidence.per_capita.0.975  <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ incidence.per_capita.fit    <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ incidence.per_capita.se.fit <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ population_unit             <dbl> 1e+05, 1e+05, 1e+05, 1e+05, 1e+05, 1e+05, …
#> $ time_unit                   <Period> 1d 0H 0M 0S, 1d 0H 0M 0S, 1d 0H 0M 0S, …