Generate a infectivity profile matrix from a long format

This converts a long format infectivity profile to something that will work with EpiEstim. In general ggoutbreak will do the conversion internally. This is provided as a utility for examples mostly.

Usage

omega_matrix(ip = i_discrete_ip, epiestim_compat = TRUE)

Arguments

ip

long format infectivity profile - a dataframe with columns:

boot (anything + default(1)) - a bootstrap identifier

probability (proportion) - the probability of new event during this period.

tau (integer + complete) - the days since the index event.

Minimally grouped by: boot (and other groupings allowed).

A default value is defined.

epiestim_compat

ensure the matrix works with EpiEstim's si_from_sample method by making probability at time zero equal to zero.

Value

a matrix with 0:max(tau) rows and 1:max(boot) columns.

Examples

omega_matrix(ggoutbreak::covid_ip)
#>           boot.1     boot.2     boot.3     boot.4     boot.5     boot.6
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.11869159 0.12523364 0.14483395 0.08783166 0.12805474 0.11393597
#>  [3,] 0.10841121 0.11495327 0.11070111 0.12168344 0.12609971 0.12429379
#>  [4,] 0.11962617 0.10654206 0.11254613 0.11985361 0.10459433 0.11770245
#>  [5,] 0.10093458 0.12803738 0.12269373 0.12991766 0.11143695 0.10640301
#>  [6,] 0.11588785 0.11308411 0.10701107 0.11802379 0.10068426 0.10922787
#>  [7,] 0.09439252 0.09532710 0.08025830 0.09881061 0.09579668 0.10169492
#>  [8,] 0.09345794 0.07850467 0.08210332 0.07136322 0.08211144 0.07250471
#>  [9,] 0.06915888 0.05887850 0.05350554 0.06129918 0.06842620 0.06591337
#> [10,] 0.04485981 0.04953271 0.05996310 0.06861848 0.04594330 0.07062147
#> [11,] 0.05514019 0.04579439 0.03966790 0.05123513 0.05669599 0.04048964
#> [12,] 0.03738318 0.03831776 0.03874539 0.03568161 0.04007820 0.03578154
#> [13,] 0.02803738 0.02336449 0.03228782 0.02195791 0.02248289 0.02730697
#> [14,] 0.01401869 0.02242991 0.01568266 0.01372370 0.01759531 0.01412429
#>           boot.7     boot.8     boot.9    boot.10    boot.11    boot.12
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.14692378 0.10066477 0.11407544 0.15413534 0.11676083 0.12580349
#>  [3,] 0.11753903 0.11965812 0.10763569 0.11184211 0.11299435 0.11019284
#>  [4,] 0.11478421 0.09876543 0.11407544 0.11184211 0.10922787 0.12029385
#>  [5,] 0.11662075 0.14339981 0.10303588 0.10244361 0.12146893 0.11019284
#>  [6,] 0.10101010 0.09971510 0.12419503 0.10526316 0.11770245 0.12304867
#>  [7,] 0.09182736 0.09116809 0.09475621 0.09022556 0.11299435 0.09641873
#>  [8,] 0.08448118 0.07977208 0.07359706 0.09116541 0.07815443 0.07070707
#>  [9,] 0.04958678 0.07787274 0.06899724 0.05451128 0.05461394 0.06519743
#> [10,] 0.06336088 0.05603039 0.06439742 0.05357143 0.05273070 0.05417815
#> [11,] 0.04683196 0.04558405 0.04691812 0.04511278 0.04237288 0.04683196
#> [12,] 0.03122130 0.03513770 0.03955842 0.04135338 0.03578154 0.03581267
#> [13,] 0.02112029 0.03798670 0.02667893 0.02631579 0.02919021 0.02479339
#> [14,] 0.01469238 0.01424501 0.02207912 0.01221805 0.01600753 0.01652893
#>          boot.13    boot.14    boot.15    boot.16    boot.17    boot.18
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.12030075 0.09412861 0.14873714 0.11428571 0.10824742 0.12011173
#>  [3,] 0.11560150 0.11835974 0.12535080 0.11244240 0.12680412 0.13221601
#>  [4,] 0.10620301 0.12767940 0.11693171 0.13179724 0.10412371 0.11824953
#>  [5,] 0.10432331 0.10438024 0.11318990 0.12626728 0.13402062 0.10148976
#>  [6,] 0.12030075 0.10717614 0.09541628 0.10783410 0.11752577 0.10707635
#>  [7,] 0.08834586 0.10344828 0.08138447 0.09400922 0.08144330 0.08193669
#>  [8,] 0.08646617 0.06710158 0.08512629 0.07649770 0.08350515 0.07821229
#>  [9,] 0.06015038 0.07362535 0.06641721 0.05898618 0.06391753 0.05865922
#> [10,] 0.05639098 0.06430568 0.05519177 0.04239631 0.04742268 0.06052142
#> [11,] 0.04887218 0.04753029 0.04396632 0.04239631 0.04639175 0.04841713
#> [12,] 0.04605263 0.04753029 0.02899906 0.05161290 0.03814433 0.04562384
#> [13,] 0.02443609 0.02795899 0.02806361 0.02764977 0.02680412 0.02700186
#> [14,] 0.02255639 0.01677540 0.01122544 0.01382488 0.02164948 0.02048417
#>          boot.19    boot.20    boot.21    boot.22    boot.23    boot.24
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.16471735 0.13674419 0.14561234 0.11764706 0.10927456 0.10472660
#>  [3,] 0.11306043 0.11162791 0.11957570 0.12246866 0.11202938 0.10750695
#>  [4,] 0.13255361 0.11162791 0.10414658 0.11089682 0.12488522 0.12604263
#>  [5,] 0.11013645 0.10511628 0.11957570 0.11475410 0.10284665 0.11862836
#>  [6,] 0.10233918 0.11162791 0.10221794 0.10414658 0.11845730 0.12604263
#>  [7,] 0.07992203 0.09395349 0.10125362 0.11282546 0.11662075 0.08989805
#>  [8,] 0.07602339 0.08186047 0.07521697 0.07907425 0.07621671 0.10009268
#>  [9,] 0.05653021 0.05395349 0.04918033 0.05593057 0.06152433 0.05746061
#> [10,] 0.05263158 0.05581395 0.05400193 0.05014465 0.06244261 0.05189991
#> [11,] 0.03606238 0.05023256 0.05014465 0.05689489 0.04499541 0.04633920
#> [12,] 0.03801170 0.04279070 0.04146577 0.03567985 0.03489440 0.03243744
#> [13,] 0.02631579 0.02511628 0.02025072 0.02410800 0.01928375 0.02687674
#> [14,] 0.01169591 0.01953488 0.01735776 0.01542912 0.01652893 0.01204819
#>          boot.25    boot.26    boot.27    boot.28    boot.29    boot.30
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.14247312 0.12727273 0.07471264 0.09878844 0.11663286 0.11678116
#>  [3,] 0.12724014 0.13014354 0.10153257 0.11090401 0.11359026 0.09322866
#>  [4,] 0.11738351 0.11866029 0.09674330 0.11929171 0.09634888 0.11776251
#>  [5,] 0.10931900 0.11578947 0.11877395 0.11929171 0.12474645 0.12953876
#>  [6,] 0.11379928 0.10622010 0.11206897 0.09972041 0.10243408 0.10696762
#>  [7,] 0.07526882 0.09473684 0.08908046 0.10531221 0.09127789 0.07850834
#>  [8,] 0.08064516 0.08133971 0.09770115 0.07362535 0.08215010 0.08439647
#>  [9,] 0.06182796 0.06028708 0.06704981 0.05964585 0.07302231 0.06280667
#> [10,] 0.05197133 0.03827751 0.06226054 0.06430568 0.06186613 0.05691855
#> [11,] 0.04211470 0.04210526 0.06417625 0.05219012 0.04563895 0.05299313
#> [12,] 0.03046595 0.03636364 0.04406130 0.04473439 0.03853955 0.05201178
#> [13,] 0.02956989 0.02870813 0.04406130 0.02982293 0.03245436 0.02944063
#> [14,] 0.01792115 0.02009569 0.02777778 0.02236719 0.02129817 0.01864573
#>          boot.31    boot.32    boot.33    boot.34    boot.35    boot.36
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.09575518 0.11560150 0.11172669 0.12912621 0.13425926 0.11797235
#>  [3,] 0.10266535 0.13251880 0.11172669 0.13786408 0.11481481 0.12073733
#>  [4,] 0.10562685 0.09116541 0.11634349 0.12135922 0.11203704 0.11152074
#>  [5,] 0.11648569 0.10338346 0.11357341 0.11456311 0.11388889 0.12258065
#>  [6,] 0.10858835 0.10056391 0.11172669 0.09320388 0.10740741 0.10230415
#>  [7,] 0.09378085 0.08364662 0.08864266 0.08543689 0.09074074 0.09953917
#>  [8,] 0.08884501 0.07800752 0.07017544 0.07766990 0.08240741 0.08110599
#>  [9,] 0.07502468 0.08740602 0.07571560 0.06407767 0.06388889 0.05437788
#> [10,] 0.05725568 0.06015038 0.06094183 0.05242718 0.05370370 0.05345622
#> [11,] 0.05725568 0.05827068 0.04986150 0.05048544 0.04074074 0.04700461
#> [12,] 0.05034551 0.03853383 0.03878116 0.04077670 0.03703704 0.04700461
#> [13,] 0.03257651 0.03383459 0.02031394 0.01941748 0.02777778 0.02949309
#> [14,] 0.01579467 0.01691729 0.03047091 0.01359223 0.02129630 0.01290323
#>          boot.37    boot.38    boot.39    boot.40    boot.41    boot.42
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.12256267 0.13339552 0.11702128 0.12740605 0.09573460 0.13994439
#>  [3,] 0.13556175 0.10354478 0.12088975 0.13107241 0.10616114 0.12233550
#>  [4,] 0.11884865 0.09888060 0.11121857 0.10724106 0.11563981 0.10843373
#>  [5,] 0.11513463 0.11847015 0.11605416 0.10540788 0.10805687 0.12696942
#>  [6,] 0.10584958 0.11287313 0.10348162 0.11090742 0.11184834 0.10009268
#>  [7,] 0.10770659 0.10074627 0.10058027 0.10724106 0.09478673 0.09545876
#>  [8,] 0.08820799 0.09421642 0.08800774 0.07607699 0.09383886 0.09916589
#>  [9,] 0.05849582 0.07182836 0.06576402 0.06782768 0.07867299 0.05189991
#> [10,] 0.05106778 0.05503731 0.04835590 0.05407883 0.04644550 0.03707136
#> [11,] 0.02971216 0.04384328 0.04448743 0.04491292 0.04360190 0.04726599
#> [12,] 0.02692665 0.03451493 0.03675048 0.04032997 0.04644550 0.03336423
#> [13,] 0.02878366 0.02145522 0.03288201 0.01558203 0.03696682 0.02502317
#> [14,] 0.01114206 0.01119403 0.01450677 0.01191567 0.02180095 0.01297498
#>          boot.43     boot.44    boot.45    boot.46    boot.47    boot.48
#>  [1,] 0.00000000 0.000000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.10714286 0.121042831 0.13594891 0.10302457 0.14811408 0.07277883
#>  [3,] 0.12500000 0.120111732 0.12682482 0.12570888 0.11591536 0.09262760
#>  [4,] 0.12030075 0.109869646 0.11770073 0.11247637 0.12419503 0.11058601
#>  [5,] 0.10244361 0.131284916 0.12043796 0.12381853 0.10671573 0.12381853
#>  [6,] 0.10808271 0.119180633 0.10492701 0.09451796 0.10671573 0.11531191
#>  [7,] 0.09304511 0.081936685 0.08211679 0.11247637 0.08831647 0.09357278
#>  [8,] 0.08176692 0.075418994 0.07116788 0.09262760 0.06991720 0.09829868
#>  [9,] 0.06672932 0.058659218 0.06660584 0.06899811 0.05979761 0.07561437
#> [10,] 0.05639098 0.063314711 0.04927007 0.05387524 0.05427783 0.06710775
#> [11,] 0.05357143 0.044692737 0.04744526 0.04725898 0.04691812 0.05954631
#> [12,] 0.03665414 0.042830540 0.04288321 0.03497164 0.03863845 0.03969754
#> [13,] 0.02537594 0.023277467 0.01824818 0.01795841 0.02023919 0.03213611
#> [14,] 0.02349624 0.008379888 0.01642336 0.01228733 0.02023919 0.01890359
#>          boot.49    boot.50    boot.51    boot.52    boot.53    boot.54
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.08679245 0.09191176 0.19630709 0.12025902 0.10211946 0.13721805
#>  [3,] 0.11981132 0.10294118 0.11953353 0.12395930 0.11271676 0.11654135
#>  [4,] 0.11603774 0.12867647 0.12050534 0.10453284 0.09730250 0.11654135
#>  [5,] 0.11792453 0.11397059 0.10689990 0.10823312 0.11946050 0.11466165
#>  [6,] 0.12075472 0.11948529 0.09815355 0.12858464 0.10308285 0.10150376
#>  [7,] 0.11509434 0.11764706 0.08066084 0.09990749 0.08381503 0.09398496
#>  [8,] 0.07452830 0.07628676 0.06413994 0.07030527 0.08766859 0.07236842
#>  [9,] 0.06226415 0.07261029 0.05733722 0.06290472 0.07803468 0.06390977
#> [10,] 0.05188679 0.05055147 0.03984451 0.05642923 0.06069364 0.05357143
#> [11,] 0.04905660 0.04411765 0.04470360 0.04532840 0.06069364 0.03947368
#> [12,] 0.03584906 0.03768382 0.03984451 0.03700278 0.04335260 0.04323308
#> [13,] 0.03113208 0.02757353 0.02040816 0.03052729 0.02793834 0.02913534
#> [14,] 0.01886792 0.01654412 0.01166181 0.01202590 0.02312139 0.01785714
#>          boot.55    boot.56    boot.57    boot.58    boot.59    boot.60
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.12896825 0.09569378 0.16431925 0.11267606 0.11254613 0.13229927
#>  [3,] 0.11706349 0.11004785 0.12394366 0.11737089 0.10977860 0.12226277
#>  [4,] 0.11805556 0.11866029 0.12394366 0.11643192 0.11070111 0.10036496
#>  [5,] 0.11408730 0.09760766 0.10516432 0.11455399 0.11439114 0.10218978
#>  [6,] 0.10317460 0.11291866 0.10798122 0.10892019 0.10885609 0.09762774
#>  [7,] 0.08730159 0.10813397 0.08544601 0.09577465 0.10147601 0.10310219
#>  [8,] 0.08928571 0.09090909 0.06384977 0.08450704 0.08118081 0.08485401
#>  [9,] 0.07440476 0.07177033 0.06291080 0.06384977 0.07011070 0.06021898
#> [10,] 0.04563492 0.05454545 0.05539906 0.05633803 0.05996310 0.06478102
#> [11,] 0.04365079 0.04976077 0.03380282 0.05070423 0.05166052 0.04470803
#> [12,] 0.04365079 0.04210526 0.03380282 0.03474178 0.03782288 0.04470803
#> [13,] 0.02281746 0.02966507 0.02723005 0.02535211 0.02121771 0.02189781
#> [14,] 0.01190476 0.01818182 0.01220657 0.01877934 0.02029520 0.02098540
#>           boot.61    boot.62    boot.63    boot.64    boot.65    boot.66
#>  [1,] 0.000000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.119744059 0.08484271 0.10334928 0.11770073 0.12686567 0.12197393
#>  [3,] 0.117915905 0.09914204 0.11578947 0.12500000 0.12500000 0.14152700
#>  [4,] 0.110603291 0.10009533 0.10717703 0.12408759 0.11007463 0.11731844
#>  [5,] 0.128884826 0.10962822 0.10047847 0.11861314 0.12593284 0.11918063
#>  [6,] 0.106946984 0.11439466 0.12344498 0.10766423 0.11380597 0.10242086
#>  [7,] 0.090493601 0.11820782 0.09569378 0.09397810 0.07555970 0.09310987
#>  [8,] 0.070383912 0.08960915 0.08421053 0.07299270 0.07649254 0.07355680
#>  [9,] 0.067641682 0.07340324 0.06411483 0.05930657 0.06250000 0.06424581
#> [10,] 0.057586837 0.05624404 0.05645933 0.05109489 0.06156716 0.04934823
#> [11,] 0.053930530 0.04766444 0.05454545 0.05018248 0.04197761 0.04003724
#> [12,] 0.033820841 0.04671115 0.04784689 0.03467153 0.03917910 0.03445065
#> [13,] 0.033820841 0.03431840 0.02679426 0.03375912 0.02705224 0.02793296
#> [14,] 0.008226691 0.02573880 0.02009569 0.01094891 0.01399254 0.01489758
#>          boot.67    boot.68    boot.69    boot.70    boot.71    boot.72
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.08366142 0.10399257 0.09181637 0.12110092 0.08695652 0.16955333
#>  [3,] 0.08661417 0.12441968 0.12475050 0.11376147 0.12422360 0.12579763
#>  [4,] 0.11811024 0.11327762 0.11776447 0.11834862 0.12008282 0.10574294
#>  [5,] 0.11811024 0.12627669 0.13473054 0.11926606 0.12215321 0.10483136
#>  [6,] 0.12007874 0.10863510 0.10379242 0.12201835 0.12111801 0.09389243
#>  [7,] 0.09645669 0.09656453 0.10479042 0.08165138 0.09834369 0.08842297
#>  [8,] 0.09055118 0.07520891 0.06287425 0.07431193 0.07142857 0.08021878
#>  [9,] 0.06988189 0.06035283 0.06586826 0.06697248 0.06107660 0.06836828
#> [10,] 0.06003937 0.04363974 0.05788423 0.05963303 0.06625259 0.04831358
#> [11,] 0.05610236 0.05663881 0.05788423 0.04311927 0.03933747 0.03828624
#> [12,] 0.05216535 0.03899721 0.03393214 0.03394495 0.05072464 0.03828624
#> [13,] 0.03051181 0.02785515 0.02994012 0.03027523 0.02070393 0.02005469
#> [14,] 0.01771654 0.02414113 0.01397206 0.01559633 0.01759834 0.01823154
#>          boot.73    boot.74    boot.75    boot.76    boot.77    boot.78 boot.79
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#>  [2,] 0.10048780 0.09192201 0.07922705 0.12335217 0.13259669 0.09824903   0.084
#>  [3,] 0.10439024 0.11884865 0.11497585 0.14030132 0.11510129 0.10214008   0.102
#>  [4,] 0.13170732 0.11977716 0.11980676 0.12052731 0.11602210 0.10797665   0.120
#>  [5,] 0.11414634 0.11977716 0.10338164 0.09981168 0.11418048 0.14105058   0.126
#>  [6,] 0.11219512 0.10677809 0.11400966 0.11205273 0.11878453 0.12062257   0.106
#>  [7,] 0.09853659 0.09935005 0.10241546 0.10451977 0.10036832 0.09630350   0.091
#>  [8,] 0.08780488 0.07892293 0.08599034 0.07532957 0.07274401 0.08365759   0.092
#>  [9,] 0.08292683 0.06592386 0.07342995 0.05743879 0.07274401 0.06420233   0.077
#> [10,] 0.04878049 0.05942433 0.04830918 0.04708098 0.04604052 0.05836576   0.060
#> [11,] 0.04878049 0.05199629 0.05410628 0.04519774 0.03775322 0.04377432   0.044
#> [12,] 0.03609756 0.04456825 0.05314010 0.03578154 0.03499079 0.03599222   0.038
#> [13,] 0.02048780 0.03249768 0.03864734 0.02542373 0.02025783 0.02723735   0.033
#> [14,] 0.01365854 0.01021356 0.01256039 0.01318267 0.01841621 0.02042802   0.027
#>          boot.80    boot.81    boot.82    boot.83    boot.84    boot.85
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.11007463 0.08812616 0.08724203 0.11688312 0.12054002 0.12147716
#>  [3,] 0.11473881 0.10946197 0.11257036 0.11317254 0.10993250 0.11175899
#>  [4,] 0.10914179 0.10111317 0.10787992 0.11410019 0.11089682 0.12536443
#>  [5,] 0.12313433 0.10482375 0.10131332 0.11502783 0.11861138 0.11175899
#>  [6,] 0.11847015 0.10575139 0.11350844 0.10667904 0.10896818 0.12147716
#>  [7,] 0.08488806 0.11317254 0.08818011 0.09183673 0.09353905 0.10106900
#>  [8,] 0.08395522 0.09183673 0.09380863 0.08256030 0.07039537 0.07385811
#>  [9,] 0.06343284 0.07606679 0.08442777 0.06864564 0.06075217 0.06511176
#> [10,] 0.05317164 0.06400742 0.05628518 0.06122449 0.05400193 0.04567541
#> [11,] 0.04570896 0.05009276 0.05440901 0.04081633 0.04918033 0.04567541
#> [12,] 0.04104478 0.03988868 0.04502814 0.04638219 0.04532305 0.03984451
#> [13,] 0.03544776 0.02968460 0.03095685 0.02319109 0.03471553 0.02137998
#> [14,] 0.01679104 0.02597403 0.02439024 0.01948052 0.02314368 0.01554908
#>          boot.86    boot.87    boot.88    boot.89    boot.90    boot.91
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.12075472 0.10432331 0.12092131 0.12465374 0.11981132 0.11338290
#>  [3,] 0.10943396 0.10244361 0.12955854 0.13481071 0.11698113 0.12360595
#>  [4,] 0.11037736 0.11560150 0.08733205 0.13296399 0.11792453 0.11895911
#>  [5,] 0.11509434 0.12312030 0.11804223 0.11449677 0.11320755 0.11152416
#>  [6,] 0.11792453 0.10714286 0.11612284 0.10156971 0.09433962 0.10315985
#>  [7,] 0.10566038 0.09398496 0.08157390 0.09695291 0.08301887 0.10315985
#>  [8,] 0.08301887 0.08834586 0.07869482 0.06094183 0.07830189 0.06691450
#>  [9,] 0.05283019 0.06954887 0.07197697 0.05724838 0.06698113 0.06412639
#> [10,] 0.06226415 0.06390977 0.05086372 0.05355494 0.06792453 0.05669145
#> [11,] 0.04622642 0.05357143 0.04990403 0.04709141 0.04339623 0.05297398
#> [12,] 0.03207547 0.04041353 0.03262956 0.03693444 0.04245283 0.04646840
#> [13,] 0.03113208 0.02631579 0.03838772 0.02862419 0.03584906 0.01765799
#> [14,] 0.01320755 0.01127820 0.02399232 0.01015697 0.01981132 0.02137546
#>          boot.92    boot.93    boot.94    boot.95    boot.96    boot.97
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.08848716 0.11985361 0.18552876 0.09356725 0.14692378 0.13395349
#>  [3,] 0.09895338 0.11985361 0.10389610 0.12670565 0.10560147 0.11813953
#>  [4,] 0.10561370 0.10064044 0.11595547 0.10136452 0.14600551 0.10976744
#>  [5,] 0.12369172 0.11436414 0.12337662 0.11890838 0.10651974 0.11255814
#>  [6,] 0.11417697 0.10430009 0.07050093 0.11598441 0.09090909 0.09581395
#>  [7,] 0.09609895 0.09332113 0.08534323 0.10916179 0.10376492 0.09674419
#>  [8,] 0.08658421 0.08234218 0.07606679 0.07894737 0.07162534 0.06697674
#>  [9,] 0.07992388 0.05672461 0.05844156 0.07699805 0.05876951 0.06697674
#> [10,] 0.06945766 0.06221409 0.06400742 0.05263158 0.05050505 0.05767442
#> [11,] 0.05518554 0.05123513 0.04081633 0.04385965 0.04224059 0.04744186
#> [12,] 0.04281637 0.04208600 0.03432282 0.04093567 0.03856749 0.04837209
#> [13,] 0.02568982 0.03385178 0.02411874 0.02436647 0.02571166 0.02976744
#> [14,] 0.01332065 0.01921317 0.01762523 0.01656920 0.01285583 0.01581395
#>          boot.98    boot.99   boot.100
#>  [1,] 0.00000000 0.00000000 0.00000000
#>  [2,] 0.12815534 0.11599626 0.08115672
#>  [3,] 0.10097087 0.11131899 0.11567164
#>  [4,] 0.11650485 0.10944808 0.10167910
#>  [5,] 0.10970874 0.12535080 0.12220149
#>  [6,] 0.09611650 0.10289991 0.11007463
#>  [7,] 0.10194175 0.08886810 0.10261194
#>  [8,] 0.07961165 0.07202993 0.08115672
#>  [9,] 0.07864078 0.06548176 0.07929104
#> [10,] 0.05631068 0.04957905 0.06250000
#> [11,] 0.03883495 0.05706268 0.05317164
#> [12,] 0.03980583 0.04396632 0.03451493
#> [13,] 0.03106796 0.04209542 0.02798507
#> [14,] 0.02233010 0.01590271 0.02798507