The Beta Distribution

Density, distribution function, quantile function and random generation for the Beta distribution with parameters shape1 and shape2 (and optional non-centrality parameter ncp).

Usage

pbeta2(q, prob, kappa, lower.tail = TRUE, log.p = FALSE)

Source

The central dbeta is based on a binomial probability, using code contributed by Catherine Loader (see dbinom) if either shape parameter is larger than one, otherwise directly from the definition. The non-central case is based on the derivation as a Poisson mixture of betas (Johnson , 1995, pp. 502–3).

The central pbeta for the default (log_p = FALSE) uses a C translation based on

Didonato, A. and Morris, A., Jr, (1992) Algorithm 708: Significant digit computation of the incomplete beta function ratios, ACM Transactions on Mathematical Software, 18, 360–373, doi:10.1145/131766.131776. (See also
Brown, B. and Lawrence Levy, L. (1994) Certification of algorithm 708: Significant digit computation of the incomplete beta, ACM Transactions on Mathematical Software, 20, 393–397, doi:10.1145/192115.192155.)
We have slightly tweaked the original “TOMS 708” algorithm, and enhanced for log.p = TRUE. For that (log-scale) case, underflow to -Inf (i.e., \(P = 0\)) or 0, (i.e., \(P = 1\)) still happens because the original algorithm was designed without log-scale considerations. Underflow to -Inf now typically signals a warning.

The non-central pbeta uses a C translation of

Lenth, R. V. (1987) Algorithm AS 226: Computing noncentral beta probabilities. Applied Statistics, 36, 241–244, doi:10.2307/2347558, incorporating
Frick, H. (1990)'s AS R84, Applied Statistics, 39, 311–2, doi:10.2307/2347780 and
Lam, M.L. (1995)'s AS R95, Applied Statistics, 44, 551–2, doi:10.2307/2986147.

This computes the lower tail only, so the upper tail suffers from cancellation and a warning will be given when this is likely to be significant.

The central case of qbeta is based on a C translation of

Cran, G. W., K. J. Martin and G. E. Thomas (1977). Remark AS R19 and Algorithm AS 109, Applied Statistics, 26, 111–114, doi:10.2307/2346887, and subsequent remarks (AS83 and correction).

Enhancements, notably for starting values and switching to a log-scale Newton search, by R Core.

The central case of rbeta is based on a C translation of

R. C. H. Cheng (1978). Generating beta variates with nonintegral shape parameters. Communications of the ACM, 21, 317–322.

Arguments

q

vector of quantiles

prob

the mean probability (vectorised)

kappa

a coefficient of variation. where 0 is no variability and 1 is maximally variability (vectorised)

lower.tail

logical; if TRUE (default), probabilities are \(P[X \le x]\), otherwise, \(P[X > x]\).

log.p

logical; if TRUE, probabilities p are given as log(p).

Value

dbeta gives the density, pbeta the distribution function, qbeta the quantile function, and rbeta generates random deviates.

Invalid arguments will result in return value NaN, with a warning.

The length of the result is determined by n for rbeta, and is the maximum of the lengths of the numerical arguments for the other functions.

The numerical arguments other than n are recycled to the length of the result. Only the first elements of the logical arguments are used.

Details

The Beta distribution with parameters shape1 \(= a\) and shape2 \(= b\) has density $$f(x)=\frac{\Gamma(a+b)}{\Gamma(a)\Gamma(b)}{x}^{a-1} {(1-x)}^{b-1}% $$ for \(a > 0\), \(b > 0\) and \(0 \le x \le 1\) where the boundary values at \(x=0\) or \(x=1\) are defined as by continuity (as limits).
The mean is \(a/(a+b)\) and the variance is \(ab/((a+b)^2 (a+b+1))\). If \(a,b > 1\), (or one of them \(=1\)), the mode is \((a-1)/(a+b-2)\). These and all other distributional properties can be defined as limits (leading to point masses at 0, 1/2, or 1) when \(a\) or \(b\) are zero or infinite, and the corresponding [dpqr]beta() functions are defined correspondingly.

pbeta is closely related to the incomplete beta function. As defined by Abramowitz and Stegun 6.6.1 $$B_x(a,b) = \int_0^x t^{a-1} (1-t)^{b-1} dt,$$ and 6.6.2 \(I_x(a,b) = B_x(a,b) / B(a,b)\) where \(B(a,b) = B_1(a,b)\) is the Beta function (beta).

\(I_x(a,b)\) is pbeta(x, a, b).

The noncentral Beta distribution (with ncp \( = \lambda\)) is defined (Johnson , 1995, pp. 502) as the distribution of \(X/(X+Y)\) where \(X \sim \chi^2_{2a}(\lambda)\) and \(Y \sim \chi^2_{2b}\). There, \(\chi^2_n(\lambda)\) is the noncentral chi-squared distribution with \(n\) degrees of freedom and non-centrality parameter \(\lambda\), see Chisquare.

Note

Supplying ncp = 0 uses the algorithm for the non-central distribution, which is not the same algorithm as when ncp is omitted. This is to give consistent behaviour in extreme cases with values of ncp very near zero.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Abramowitz, M. and Stegun, I. A. (1972) Handbook of Mathematical Functions. New York: Dover. Chapter 6: Gamma and Related Functions.

Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, volume 2, especially chapter 25. Wiley, New York.

See also

Examples

pbeta2(c(0.25,0.5,0.75), 0.5, 0.25)
#> [1] 0.0001270637 0.5000000000 0.9998729363