Plot an incidence or proportion versus growth phase diagram

Plot an incidence or proportion versus growth phase diagram

Usage

plot_growth_phase(
  modelled,
  timepoints = NULL,
  duration = max(dplyr::count(modelled)$n),
  interval = 7,
  mapping = if (interfacer::is_col_present(modelled, class)) ggplot2::aes(colour = class)
    else ggplot2::aes(),
  cis = TRUE,
  ...
)

Arguments

modelled

Growth rates and incidence / proportion timeseries as the outputs of functions such as proportion_locfit_model, poisson_locfit_model, or similar. - EITHER: a dataframe with columns:

time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`

incidence.per_capita.fit (double) - an estimate of the incidence per capita rate on a log scale

incidence.per_capita.se.fit (positive_double) - the standard error of the incidence per capita rate estimate on a log scale

incidence.per_capita.0.025 (positive_double) - lower confidence limit of the incidence per capita rate (true scale)

incidence.per_capita.0.5 (positive_double) - median estimate of the incidence per capita rate (true scale)

incidence.per_capita.0.975 (positive_double) - upper confidence limit of the incidence per capita rate (true scale)

population_unit (double) - The population unit on which the per capita incidence rate is calculated

time_unit (lubridate::as.period) - The time period over which the per capita incidence rate is calculated

growth.fit (double) - an estimate of the growth rate

growth.se.fit (positive_double) - the standard error the growth rate

growth.0.025 (double) - lower confidence limit of the growth rate

growth.0.5 (double) - median estimate of the growth rate

growth.0.975 (double) - upper confidence limit of the growth rate

Any grouping allowed.

OR with columns:

time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`

incidence.fit (double) - an estimate of the incidence rate on a log scale

incidence.se.fit (positive_double) - the standard error of the incidence rate estimate on a log scale

incidence.0.025 (positive_double) - lower confidence limit of the incidence rate (true scale)

incidence.0.5 (positive_double) - median estimate of the incidence rate (true scale)

incidence.0.975 (positive_double) - upper confidence limit of the incidence rate (true scale)

growth.fit (double) - an estimate of the growth rate

growth.se.fit (positive_double) - the standard error the growth rate

growth.0.025 (double) - lower confidence limit of the growth rate

growth.0.5 (double) - median estimate of the growth rate

growth.0.975 (double) - upper confidence limit of the growth rate

Any grouping allowed.

OR with columns:

time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`

risk_ratio.0.025 (positive_double) - lower confidence limit of the excess risk ratio for a population group

risk_ratio.0.5 (positive_double) - median estimate of the excess risk ratio for a population group

risk_ratio.0.975 (positive_double) - upper confidence limit of the excess risk ratio for a population group

proportion.fit (double) - an estimate of the proportion on a logit scale

proportion.se.fit (positive_double) - the standard error of proportion estimate on a logit scale

proportion.0.025 (proportion) - lower confidence limit of proportion (true scale)

proportion.0.5 (proportion) - median estimate of proportion (true scale)

proportion.0.975 (proportion) - upper confidence limit of proportion (true scale)

relative.growth.fit (double) - an estimate of the relative growth rate

relative.growth.se.fit (positive_double) - the standard error the relative growth rate

relative.growth.0.025 (double) - lower confidence limit of the relative growth rate

relative.growth.0.5 (double) - median estimate of the relative growth rate

relative.growth.0.975 (double) - upper confidence limit of the relative growth rate

Any grouping allowed.

OR with columns:

time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`

proportion.fit (double) - an estimate of the proportion on a logit scale

proportion.se.fit (positive_double) - the standard error of proportion estimate on a logit scale

proportion.0.025 (proportion) - lower confidence limit of proportion (true scale)

proportion.0.5 (proportion) - median estimate of proportion (true scale)

proportion.0.975 (proportion) - upper confidence limit of proportion (true scale)

relative.growth.fit (double) - an estimate of the relative growth rate

relative.growth.se.fit (positive_double) - the standard error the relative growth rate

relative.growth.0.025 (double) - lower confidence limit of the relative growth rate

relative.growth.0.5 (double) - median estimate of the relative growth rate

relative.growth.0.975 (double) - upper confidence limit of the relative growth rate

Any grouping allowed.

OR with NULL

timepoints

time points (as Date or time_period vector) of dates to plot phase diagrams. If multiple this will result in a sequence of plots as facets. If NULL (the default) it will be the last time point in the series

duration

the length of the growth rate phase trail

interval

the length of time between markers on the phase plot

mapping

a ggplot2::aes() mapping

cis

logical; should the phases be marked with confidence intervals?

...

Named arguments passed on to geom_events

events

Significant events or time spans - a dataframe with columns:

label (character) - the event label

start (date) - the start date, or the date of the event

end (date) - the end date or NA if a single event

Any grouping allowed.

A default value is defined.

event_label_size

how big to make the event label

event_label_colour

the event label colour

event_label_angle

the event label colour

event_line_colour

the event line colour

event_fill_colour

the event area fill

hide_labels

do not show labels at all

guide_axis

a guide axis configuration for the labels (see ggplot2::guide_axis and ggplot2::dup_axis). This can be used to specify a position amongst other things.

...

Named arguments passed on to ggplot2::scale_x_date

name

The name of the scale. Used as the axis or legend title. If waiver(), the default, the name of the scale is taken from the first mapping used for that aesthetic. If NULL, the legend title will be omitted.

breaks

One of:

NULL for no breaks

waiver() for the breaks specified by date_breaks

A Date/POSIXct vector giving positions of breaks

A function that takes the limits as input and returns breaks as output

date_breaks

A string giving the distance between breaks like "2 weeks", or "10 years". If both breaks and date_breaks are specified, date_breaks wins. Valid specifications are 'sec', 'min', 'hour', 'day', 'week', 'month' or 'year', optionally followed by 's'.

labels

One of:

NULL for no labels

waiver() for the default labels computed by the transformation object

A character vector giving labels (must be same length as breaks)

An expression vector (must be the same length as breaks). See ?plotmath for details.

A function that takes the breaks as input and returns labels as output. Also accepts rlang lambda function notation.

date_labels

A string giving the formatting specification for the labels. Codes are defined in strftime(). If both labels and date_labels are specified, date_labels wins.

minor_breaks

One of:

NULL for no breaks

waiver() for the breaks specified by date_minor_breaks

A Date/POSIXct vector giving positions of minor breaks

A function that takes the limits as input and returns minor breaks as output

date_minor_breaks

A string giving the distance between minor breaks like "2 weeks", or "10 years". If both minor_breaks and date_minor_breaks are specified, date_minor_breaks wins. Valid specifications are 'sec', 'min', 'hour', 'day', 'week', 'month' or 'year', optionally followed by 's'.

limits

One of:

NULL to use the default scale range

A numeric vector of length two providing limits of the scale. Use NA to refer to the existing minimum or maximum

A function that accepts the existing (automatic) limits and returns new limits. Also accepts rlang lambda function notation. Note that setting limits on positional scales will remove data outside of the limits. If the purpose is to zoom, use the limit argument in the coordinate system (see coord_cartesian()).

expand

For position scales, a vector of range expansion constants used to add some padding around the data to ensure that they are placed some distance away from the axes. Use the convenience function expansion() to generate the values for the expand argument. The defaults are to expand the scale by 5% on each side for continuous variables, and by 0.6 units on each side for discrete variables.

oob

One of:

Function that handles limits outside of the scale limits (out of bounds). Also accepts rlang lambda function notation.

The default (scales::censor()) replaces out of bounds values with NA.

scales::squish() for squishing out of bounds values into range.

scales::squish_infinite() for squishing infinite values into range.

guide

A function used to create a guide or its name. See guides() for more information.

position

For position scales, The position of the axis. left or right for y axes, top or bottom for x axes.

sec.axis

sec_axis() is used to specify a secondary axis.

timezone

The timezone to use for display on the axes. The default (NULL) uses the timezone encoded in the data.

na.value

Missing values will be replaced with this value.

Named arguments passed on to ggplot2::facet_wrap

facets

A set of variables or expressions quoted by vars() and defining faceting groups on the rows or columns dimension. The variables can be named (the names are passed to labeller).

For compatibility with the classic interface, can also be a formula or character vector. Use either a one sided formula, ~a + b, or a character vector, c("a", "b").

nrow,ncol

Number of rows and columns.

scales

Should scales be fixed ("fixed", the default), free ("free"), or free in one dimension ("free_x", "free_y")?

shrink

If TRUE, will shrink scales to fit output of statistics, not raw data. If FALSE, will be range of raw data before statistical summary.

labeller

A function that takes one data frame of labels and returns a list or data frame of character vectors. Each input column corresponds to one factor. Thus there will be more than one with vars(cyl, am). Each output column gets displayed as one separate line in the strip label. This function should inherit from the "labeller" S3 class for compatibility with labeller(). You can use different labeling functions for different kind of labels, for example use label_parsed() for formatting facet labels. label_value() is used by default, check it for more details and pointers to other options.

as.table

If TRUE, the default, the facets are laid out like a table with highest values at the bottom-right. If FALSE, the facets are laid out like a plot with the highest value at the top-right.

switch

By default, the labels are displayed on the top and right of the plot. If "x", the top labels will be displayed to the bottom. If "y", the right-hand side labels will be displayed to the left. Can also be set to "both".

drop

If TRUE, the default, all factor levels not used in the data will automatically be dropped. If FALSE, all factor levels will be shown, regardless of whether or not they appear in the data.

dir

Direction: either "h" for horizontal, the default, or "v", for vertical.

strip.position

By default, the labels are displayed on the top of the plot. Using strip.position it is possible to place the labels on either of the four sides by setting strip.position = c("top", "bottom", "left", "right")

axes

Determines which axes will be drawn in case of fixed scales. When "margins" (default), axes will be drawn at the exterior margins. "all_x" and "all_y" will draw the respective axes at the interior panels too, whereas "all" will draw all axes at all panels.

axis.labels

Determines whether to draw labels for interior axes when the scale is fixed and the axis argument is not "margins". When "all" (default), all interior axes get labels. When "margins", only the exterior axes get labels, and the interior axes get none. When "all_x" or "all_y", only draws the labels at the interior axes in the x- or y-direction respectively.

Value

a ggplot

Examples

tmp = ggoutbreak::england_covid %>%
  time_aggregate(count=sum(count))

tmp_pop = ggoutbreak::england_demographics %>%
  dplyr::ungroup() %>%
  dplyr::summarise(population = sum(population))

# If the incidence is normalised by population
tmp2 = tmp %>%
  poisson_locfit_model() %>%
  normalise_incidence(tmp_pop)

timepoints = as.Date(c("Lockdown 1" = "2020-03-30", "Lockdown 2" = "2020-12-31"))

if(interactive()) {
  plot_growth_phase(tmp2, timepoints, duration=108)
}