Poisson time-series model.

Takes a list of times and counts and fits a quasi-poisson model fitted with a log link function to count data using local regression using the package locfit.

Usage

poisson_locfit_model(
  d = i_incidence_input,
  ...,
  window = 14,
  deg = 1,
  frequency = "1 day",
  predict = TRUE,
  .progress = interactive()
)

Arguments

d

input data - a dataframe with columns:

count (positive_integer) - Positive case counts associated with the specified time frame

time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`

Ungrouped.

...

not used and present to allow proportion model to be used in a group_modify

window

a number of data points defining the bandwidth of the estimate, smaller values result in less smoothing, large value in more. The default value of 14 is calibrated for data provided on a daily frequency, with weekly data a lower value may be preferred. - default 14

deg

polynomial degree (min 1) - higher degree results in less smoothing, lower values result in more smoothing. A degree of 1 is fitting a linear model piece wise. - default 1

frequency

the density of the output estimates as a time period such as 7 days or 2 weeks. - default 1 day

predict

result is a prediction dataframe. If false we return the locfit models (advanced). - default TRUE

.progress

show a CLI progress bar

Value

A dataframe containing the following columns:

time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a time_period

incidence.fit (double) - an estimate of the incidence rate on a log scale

incidence.se.fit (positive_double) - the standard error of the incidence rate estimate on a log scale

incidence.0.025 (positive_double) - lower confidence limit of the incidence rate (true scale)

incidence.0.5 (positive_double) - median estimate of the incidence rate (true scale)

incidence.0.975 (positive_double) - upper confidence limit of the incidence rate (true scale)

growth.fit (double) - an estimate of the growth rate

growth.se.fit (positive_double) - the standard error the growth rate

growth.0.025 (double) - lower confidence limit of the growth rate

growth.0.5 (double) - median estimate of the growth rate

growth.0.975 (double) - upper confidence limit of the growth rate

Any grouping allowed.

Details

This results is an incidence rate estimate plus an absolute exponential growth rate estimate both based on the time unit of the input data (e.g. for daily data the rate will be cases per day and the growth rate will be daily).

Examples

withr::with_options(list("ggoutbreak.keep_cdf"=TRUE),{
  ggoutbreak::england_covid %>%
    ggoutbreak::poisson_locfit_model(window=21) %>%
    dplyr::glimpse()
})
#> Rows: 26,790
#> Columns: 22
#> Groups: class [19]
#> $ class            <fct> 00_04, 00_04, 00_04, 00_04, 00_04, 00_04, 00_04, 00_0…
#> $ time             <time_prd> 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14…
#> $ incidence.fit    <dbl> -17.891227, -17.395804, -16.873590, -16.331042, -15.7…
#> $ incidence.se.fit <dbl> 1.9111699, 1.9238465, 1.9194060, 1.9000358, 1.8681012…
#> $ incidence.0.025  <dbl> 4.010032e-10, 6.419738e-10, 1.091671e-09, 1.950766e-0…
#> $ incidence.0.05   <dbl> 7.323064e-10, 1.177056e-09, 1.998775e-09, 3.549985e-0…
#> $ incidence.0.25   <dbl> 4.678483e-09, 7.612924e-09, 1.287203e-08, 2.243606e-0…
#> $ incidence.0.5    <dbl> 1.698004e-08, 2.786752e-08, 4.697785e-08, 8.081996e-0…
#> $ incidence.0.75   <dbl> 6.162718e-08, 1.020106e-07, 1.714507e-07, 2.911325e-0…
#> $ incidence.0.95   <dbl> 3.937173e-07, 6.597809e-07, 1.104136e-06, 1.839970e-0…
#> $ incidence.0.975  <dbl> 7.190010e-07, 1.209705e-06, 2.021596e-06, 3.348359e-0…
#> $ incidence.cdf    <list> function (x) , {,     fn(trans_fn(x), p1, p2), }, fu…
#> $ growth.fit       <dbl> 0.4798757, 0.4792888, 0.4776793, 0.4752744, 0.4723010…
#> $ growth.se.fit    <dbl> 0.05395200, 0.05584978, 0.05729669, 0.05835042, 0.059…
#> $ growth.0.025     <dbl> 0.3741317, 0.3698252, 0.3653799, 0.3609097, 0.3565117…
#> $ growth.0.05      <dbl> 0.3911325, 0.3874241, 0.3834346, 0.3792965, 0.3751275…
#> $ growth.0.25      <dbl> 0.4434856, 0.4416187, 0.4390333, 0.4359176, 0.4324540…
#> $ growth.0.5       <dbl> 0.4798757, 0.4792888, 0.4776793, 0.4752744, 0.4723010…
#> $ growth.0.75      <dbl> 0.5162658, 0.5169589, 0.5163253, 0.5146311, 0.5121480…
#> $ growth.0.95      <dbl> 0.5686188, 0.5711535, 0.5719240, 0.5712523, 0.5694745…
#> $ growth.0.975     <dbl> 0.5856197, 0.5887523, 0.5899788, 0.5896391, 0.5880903…
#> $ growth.cdf       <list> function (x) , {,     fn(trans_fn(x), p1, p2), }, fu…