Calculate a normalised risk ratio from proportions

This assumes that for example, case distribution proportions are stratified by a population grouping, e.g. geography or age, and we have estimates of the size of that population during that time period. Normalising by population proportion allows us to compare the relative risk of the outcome in groups, compared to the expected population risk if the outcome is evenly distributed across the population. There may be other proportions other than population fractions that may be useful to compare. At the moment this does not handle any uncertainty.

Usage

infer_risk_ratio(
  modelled = i_proportion_model,
  base = i_baseline_proportion_data,
  ...
)

Arguments

modelled

Model output from something like proportion_locfit_model() - a dataframe with columns:

time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`

proportion.fit (double) - an estimate of the proportion on a logit scale

proportion.se.fit (positive_double) - the standard error of proportion estimate on a logit scale

proportion.0.025 (proportion) - lower confidence limit of proportion (true scale)

proportion.0.5 (proportion) - median estimate of proportion (true scale)

proportion.0.975 (proportion) - upper confidence limit of proportion (true scale)

Any grouping allowed.

base

The baseline data must be grouped in the same way as modelled. It may be a time series but does not have to be. - a dataframe with columns:

baseline_proportion (proportion) - Baseline proportion for comparison

Any grouping allowed.

...

not used

Value

a dataframe with relative risk / risk ratio columns. A dataframe containing the following columns:

time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a time_period

risk_ratio.0.025 (positive_double) - lower confidence limit of the excess risk ratio for a population group

risk_ratio.0.5 (positive_double) - median estimate of the excess risk ratio for a population group

risk_ratio.0.975 (positive_double) - upper confidence limit of the excess risk ratio for a population group

Any grouping allowed.

Examples

tmp = ggoutbreak::england_covid_proportion_age_stratified  %>%
  ggoutbreak::infer_risk_ratio(ggoutbreak::england_demographics) %>%
  dplyr::glimpse()
#> Rows: 26,790
#> Columns: 24
#> Groups: class [19]
#> $ class                  <fct> 00_04, 00_04, 00_04, 00_04, 00_04, 00_04, 00_04…
#> $ time                   <time_prd> 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, …
#> $ proportion.fit         <dbl> -1000.00000, -1000.00000, -1000.00000, -1000.00…
#> $ proportion.se.fit      <dbl> 0.01129143, 0.01127551, 0.01125958, 0.01124366,…
#> $ proportion.0.025       <dbl> 0.000000e+00, 0.000000e+00, 0.000000e+00, 0.000…
#> $ proportion.0.05        <dbl> 0.000000e+00, 0.000000e+00, 0.000000e+00, 0.000…
#> $ proportion.0.25        <dbl> 0.000000e+00, 0.000000e+00, 0.000000e+00, 0.000…
#> $ proportion.0.5         <dbl> 0.000000e+00, 0.000000e+00, 0.000000e+00, 0.000…
#> $ proportion.0.75        <dbl> 0.000000e+00, 0.000000e+00, 0.000000e+00, 0.000…
#> $ proportion.0.95        <dbl> 0.000000e+00, 0.000000e+00, 0.000000e+00, 0.000…
#> $ proportion.0.975       <dbl> 0.000000e+00, 0.000000e+00, 0.000000e+00, 0.000…
#> $ relative.growth.fit    <dbl> 0.00000000, 0.00000000, 0.00000000, 0.00000000,…
#> $ relative.growth.se.fit <dbl> 1.626352e-05, 1.626352e-05, 1.626352e-05, 1.626…
#> $ relative.growth.0.025  <dbl> -3.187591e-05, -3.187591e-05, -3.187591e-05, -3…
#> $ relative.growth.0.05   <dbl> -2.675111e-05, -2.675111e-05, -2.675111e-05, -2…
#> $ relative.growth.0.25   <dbl> -1.096958e-05, -1.096958e-05, -1.096958e-05, -1…
#> $ relative.growth.0.5    <dbl> 0.00000000, 0.00000000, 0.00000000, 0.00000000,…
#> $ relative.growth.0.75   <dbl> 1.096958e-05, 1.096958e-05, 1.096958e-05, 1.096…
#> $ relative.growth.0.95   <dbl> 2.675111e-05, 2.675111e-05, 2.675111e-05, 2.675…
#> $ relative.growth.0.975  <dbl> 3.187591e-05, 3.187591e-05, 3.187591e-05, 3.187…
#> $ baseline_proportion    <dbl> 0.05447011, 0.05447011, 0.05447011, 0.05447011,…
#> $ risk_ratio.0.025       <dbl> 0.000000e+00, 0.000000e+00, 0.000000e+00, 0.000…
#> $ risk_ratio.0.5         <dbl> 0.000000e+00, 0.000000e+00, 0.000000e+00, 0.000…
#> $ risk_ratio.0.975       <dbl> 0.000000e+00, 0.000000e+00, 0.000000e+00, 0.000…

if(interactive()) {
  plot_growth_phase(tmp,duration = 14*4)+
  ggplot2::scale_colour_viridis_d()+
  ggplot2::coord_cartesian(xlim=c(-0.25,0.25),ylim=c(0.02,20))
}