Rescale a timeseries in the temporal dimension

Sometimes we may have, for example, modelled incidence or growth rates on weekly data resulting in cases per week and growth rate per week. We may wish to use this to estimate the reproduction number, using algorithms that assume a daily incidence. Not everything has a dependence on time, and things such as proportions, or prevalence will not change.

Usage

rescale_model(df = i_timeseries, time_unit)

Arguments

df

A data frame containing modelled output. This will modify the following columns if present:

A dataframe containing the following columns:

time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a time_period

incidence.fit (double) - an estimate of the incidence rate on a log scale

incidence.se.fit (positive_double) - the standard error of the incidence rate estimate on a log scale

incidence.0.025 (positive_double) - lower confidence limit of the incidence rate (true scale)

incidence.0.5 (positive_double) - median estimate of the incidence rate (true scale)

incidence.0.975 (positive_double) - upper confidence limit of the incidence rate (true scale)

Any grouping allowed.

A dataframe containing the following columns:

time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a time_period

growth.fit (double) - an estimate of the growth rate

growth.se.fit (positive_double) - the standard error the growth rate

growth.0.025 (double) - lower confidence limit of the growth rate

growth.0.5 (double) - median estimate of the growth rate

growth.0.975 (double) - upper confidence limit of the growth rate

Any grouping allowed.

A dataframe containing the following columns:

time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a time_period

proportion.fit (double) - an estimate of the proportion on a logit scale

proportion.se.fit (positive_double) - the standard error of proportion estimate on a logit scale

proportion.0.025 (proportion) - lower confidence limit of proportion (true scale)

proportion.0.5 (proportion) - median estimate of proportion (true scale)

proportion.0.975 (proportion) - upper confidence limit of proportion (true scale)

relative.growth.fit (double) - an estimate of the relative growth rate

relative.growth.se.fit (positive_double) - the standard error the relative growth rate

relative.growth.0.025 (double) - lower confidence limit of the relative growth rate

relative.growth.0.5 (double) - median estimate of the relative growth rate

relative.growth.0.975 (double) - upper confidence limit of the relative growth rate

Any grouping allowed.

time_unit

a lubridate period string such as "1 day"

Value

the same time series with different time unit, and adjusted incidence and growth rate figures.

Examples

sim = sim_poisson_model(time_unit = "1 week")
#> interfacer: development mode active (local function).
#> This message is displayed once every 8 hours.
incidence = sim %>% poisson_locfit_model(frequency = "1 day", deg = 2, window=5)
incidence2 = incidence %>% rescale_model(time_unit = "1 day")
incidence2 %>% dplyr::glimpse()
#> Rows: 727
#> Columns: 20
#> Groups: statistic [1]
#> $ statistic        <chr> "infections", "infections", "infections", "infections…
#> $ time             <time_prd> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 1…
#> $ incidence.fit    <dbl> 2.496489, 2.517828, 2.539042, 2.560129, 2.581089, 2.6…
#> $ incidence.se.fit <dbl> 0.04338465, 0.04060679, 0.03808768, 0.03581637, 0.033…
#> $ incidence.0.025  <dbl> 11.15019, 11.45287, 11.75632, 12.06043, 12.36510, 12.…
#> $ incidence.0.05   <dbl> 79.12566, 81.20251, 83.28791, 85.38124, 87.48201, 89.…
#> $ incidence.0.25   <dbl> 82.52788, 84.46604, 86.42373, 88.40082, 90.39724, 92.…
#> $ incidence.0.5    <dbl> 12.13979, 12.40163, 12.66753, 12.93749, 13.21152, 13.…
#> $ incidence.0.75   <dbl> 87.50196, 89.22197, 90.98019, 92.77682, 94.61203, 96.…
#> $ incidence.0.95   <dbl> 91.26434, 92.80780, 94.40563, 96.05796, 97.76485, 99.…
#> $ incidence.0.975  <dbl> 13.21722, 13.42899, 13.64936, 13.87832, 14.11587, 14.…
#> $ growth.fit       <dbl> 0.02140392, 0.02127503, 0.02114210, 0.02100573, 0.020…
#> $ growth.se.fit    <dbl> 0.003824766, 0.003690481, 0.003554655, 0.003417722, 0…
#> $ growth.0.025     <dbl> 0.01390752, 0.01404182, 0.01417510, 0.01430712, 0.014…
#> $ growth.0.05      <dbl> 0.1057892, 0.1064331, 0.1070665, 0.1076885, 0.1082984…
#> $ growth.0.25      <dbl> 0.1317691, 0.1315008, 0.1312116, 0.1309036, 0.1305787…
#> $ growth.0.5       <dbl> 0.02140392, 0.02127503, 0.02114210, 0.02100573, 0.020…
#> $ growth.0.75      <dbl> 0.1678858, 0.1663495, 0.1647777, 0.1631766, 0.1615524…
#> $ growth.0.95      <dbl> 0.1938657, 0.1914173, 0.1889229, 0.1863917, 0.1838328…
#> $ growth.0.975     <dbl> 0.02890033, 0.02850824, 0.02810909, 0.02770434, 0.027…